cleanup: refactor retry to InstanceAdminConnection
This removes a number of places where we created a `grpc::ClientContext`
that went unused and a number of functions that ignored their
`grpc::ClientContext` too.
