+++ google/cloud/spanner/spanner_client.bzl
@@ -39,7 +39,6 @@ spanner_client_hdrs = [
     "internal/database_admin_retry.h",
     "internal/database_admin_stub.h",
     "internal/date.h",
-    "internal/instance_admin_retry.h",
     "internal/instance_admin_stub.h",
     "internal/log_wrapper.h",
     "internal/logging_spanner_stub.h",
@@ -92,7 +91,6 @@ spanner_client_srcs = [
     "internal/database_admin_retry.cc",
     "internal/database_admin_stub.cc",
     "internal/date.cc",
-    "internal/instance_admin_retry.cc",
     "internal/instance_admin_stub.cc",
     "internal/logging_spanner_stub.cc",
     "internal/merge_chunk.cc",
